Added Listmods
Listmods shows users by usernames and discriminators (because user mentions are not supported in embeds), unless the user is no longer in the server; Then it shows by saying 'deleted-user []' in place of username/discrim combo. Shows moderator roles as mentions (because they are supported in embeds to my knowledge) unless the role cannot be found; Then it shows as 'deleted-role []'. Embed color is based on highest colored role or configs/template.embed.colors.help